% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eea_plot.R
\name{eea_plot}
\alias{eea_plot}
\title{Funkcja wyswietla wykres na podstawie wyniku funkcji "eea_szaco_emisji"}
\usage{
eea_plot(wynik = out)
}
\arguments{
\item{wynik}{\itemize{
\item data frame, jest to wynik funkcji "eea_szaco_emisji"
}}
}
\value{
plot
}
\description{
Przyjmuje nastepujace argumenty:
}
