% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eea_plot.R, R/eea_szaco_emisji.R
\name{eea_szaco_emisji}
\alias{eea_szaco_emisji}
\title{Funkcja wyswietla wykres na podstawie wyniku funkcji "eea_szaco_emisji"}
\usage{
eea_szaco_emisji(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("EC", "CO")
)

eea_szaco_emisji(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("EC", "CO")
)
}
\arguments{
\item{dane}{\itemize{
\item data frame, sklada sie z 50 wierszy o kolumnach: (Nat, segment, paliwo, technologia)
\item nat, to liczba losowo generowana
\item segment, character w tej kolumnie losowane sa wartosci sposrod 4 dostepnych
parametrów: "Mini", "Small", "Medium", "Large-SUV-Executive"
\item paliwo, okresla sie rodzaj paliwa
\item technologia, okresla rodzaj technologii oczyszczania spalin
}}

\item{kategoria}{\itemize{
\item character okresla kategorie pojazdu
}}

\item{euro}{\itemize{
\item Europejski standard emisji spalin
}}

\item{mode}{\itemize{
\item zawiera rodzaj drogi
}}

\item{substancja}{\itemize{
\item wektor character, przyjmuje substancje dla ktorych chcemy szacowac emisje
aby zobaczyc liste dostepnych wartosci dla poszczegolnych parametrow nalezy uzyc
funkcji unique dla danych "wskazniki"
np. unique(wskazniki$Category)
}}

\item{wynik}{\itemize{
\item data frame, jest to wynik funkcji "eea_szaco_emisji"
}}
}
\value{
plot

data frame z obliczona emisja
}
\description{
Przyjmuje nastepujace argumenty:

Przyjmuje nastpujepujace argumenty:
}
