% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeaemisje.R
\name{eeaemisje}
\alias{eeaemisje}
\title{eeaemisje: pakiet do pracy z danymi z emisji pobranymi ze strony EEA}
\description{
Pakiet sklada sie z 2 funkcji. Do liczenia szacowanej emisji substancji "eea_szaco_emisji" oraz
do wizualizacji wyniku tej funkcji "eea_plot". Aby uzyskać pomoc do konkretnej funckji wpisz- ?"eea_szaco_emisji"
}
