% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pack.R
\name{fun_pack}
\alias{fun_pack}
\title{Funkcja liczaca emisje}
\usage{
fun_pack(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("EC", "CO")
)
}
\arguments{
\item{substancja}{}
}
\value{
data frame z obliczona emisja
}
\description{
Funkcja liczaca emisje
}
\examples{

}
