% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeaemisje.R
\name{eea_emisje}
\alias{eea_emisje}
\title{eea_emisje: pakiet do pracy z danymi z emisji pobranymi ze strony EEA}
\description{
Pakiet sklada sie z 2 funkcji. Do liczenia szacowanej emisji substancji oraz do wizualizacji.
}
